refac: First Step in refactoring.
Fakemon Data is now calculated independently of the Framework. Next step: reduce number of Components.

The goal of this refactoring is to make the domain independent of the frontend. This will allow better control of the generated data (more sensible suggestions, that take into account all available data) and as an added Bonus, a reroll can be done without reloading the page.